# dev-helpers/swagger.json
{"openapi":"3.0.0","paths":{"/assets-cdn/{network}/accounts":{"get":{"operationId":"AccountsController_getAllAccountAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}}],"responses":{"200":{"description":""}},"tags":["accounts"]}},"/assets-cdn/{network}/accounts/{address}":{"get":{"operationId":"AccountsController_getAccountAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["accounts"]}},"/assets-cdn/{network}/accounts/{address}/icon.png":{"get":{"operationId":"AccountsController_getAccountIconPng","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["accounts"]}},"/assets-cdn/{network}/accounts/{address}/icon.svg":{"get":{"operationId":"AccountsController_getAccountIconSvg","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["accounts"]}},"/assets-cdn/{network}/identities":{"get":{"operationId":"IdentitiesController_getAllIdentityAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}}],"responses":{"200":{"description":""}},"tags":["identities"]}},"/assets-cdn/{network}/identities/{identity}":{"get":{"operationId":"IdentitiesController_getIdentityAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"identity","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["identities"]}},"/assets-cdn/{network}/identities/{identity}/icon.png":{"get":{"operationId":"IdentitiesController_getIdentityLogo","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"identity","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["identities"]}},"/assets-cdn/{network}/tokens":{"get":{"operationId":"TokensController_getAllTokenAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}}],"responses":{"200":{"description":""}},"tags":["tokens"]}},"/assets-cdn/{network}/tokens/{identifier}":{"get":{"operationId":"TokensController_getTokenAssets","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tokens"]}},"/assets-cdn/{network}/tokens/{identifier}/icon.png":{"get":{"operationId":"TokensController_getTokenLogoPng","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tokens"]}},"/assets-cdn/{network}/tokens/{identifier}/icon.svg":{"get":{"operationId":"TokensController_getTokenLogoSvg","parameters":[{"name":"network","required":false,"in":"path","description":"The network to get identities from. Default is MAINNET","schema":{"type":"string","enum":["mainnet","testnet","devnet"]}},{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tokens"]}}},"info":{"title":"MultiversX Assets CDN","description":"## Welcome to the MultiversX Microservice API!\n\nHere you can set your custom documentation in markdown format\n","version":"1.0.0"},"externalDocs":{"description":"MultiversX Docs","url":"https://docs.multiversx.com/"},"components":{"schemas":{"AccountAssetsSocial":{"type":"object","properties":{"website":{"type":"string"},"email":{"type":"string"},"blog":{"type":"string"},"twitter":{"type":"string"},"discord":{"type":"string"},"telegram":{"type":"string"},"facebook":{"type":"string"},"instagram":{"type":"string"},"youtube":{"type":"string"},"whitepaper":{"type":"string"},"coinmarketcap":{"type":"string"},"coingecko":{"type":"string"},"linkedin":{"type":"string"}},"required":["website","email","blog","twitter","discord","telegram","facebook","instagram","youtube","whitepaper","coinmarketcap","coingecko","linkedin"]},"AccountAssets":{"type":"object","properties":{"address":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"social":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/AccountAssetsSocial"}]},"tags":{"type":"array","items":{"type":"string"}},"proof":{"type":"string","nullable":true},"iconPng":{"type":"string","nullable":true},"iconSvg":{"type":"string","nullable":true}},"required":["address","name","description","social","tags","proof","iconPng","iconSvg"]},"IdentityAssets":{"type":"object","properties":{"identity":{"type":"string"},"name":{"type":"string"},"owners":{"type":"array","items":{"type":"string"}},"description":{"type":"string","nullable":true},"avatar":{"type":"string","nullable":true},"twitter":{"type":"string","nullable":true},"website":{"type":"string","nullable":true},"location":{"type":"string","nullable":true}},"required":["identity","name","owners","description","avatar","twitter","website","location"]},"NftRank":{"type":"object","properties":{"identifier":{"type":"string"},"rank":{"type":"number"}},"required":["identifier","rank"]},"TokenAssets":{"type":"object","properties":{"identifier":{"type":"string"},"website":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["active","inactive"],"default":"inactive"},"name":{"type":"string"},"pngUrl":{"type":"string"},"svgUrl":{"type":"string"},"ledgerSignature":{"type":"string","nullable":true},"lockedAccounts":{"type":"string","nullable":true},"extraTokens":{"nullable":true,"type":"array","items":{"type":"string"}},"preferredRankAlgorithm":{"type":"string","enum":["trait","statistical","openRarity","jaccardDistances","custom"],"nullable":true,"description":"Preferred ranking algorithm for NFT collections. Supported values are \"trait\", \"statistical\", \"jaccardDistances\", \"openRarity\" and \"custom\"."},"priceSource":{"type":"number","enum":[],"nullable":true},"ranks":{"nullable":true,"type":"array","items":{"$ref":"#/components/schemas/NftRank"}},"social":{"type":"object","nullable":true}},"required":["identifier","website","description","status","name","pngUrl","svgUrl","ledgerSignature","lockedAccounts","extraTokens","preferredRankAlgorithm","priceSource","ranks","social"]}}}}